created base class to clean up methods
